refactor(slides): extract deck op engine to services
applyDeckOps + the op zod schemas (update/insert/move/delete/reorder/
set_theme), DeckOpError, and findSlide move VERBATIM from the MCP deck tool
into packages/services/src/slides/deckOps.ts so the editor save path can
share one engine. deck.ts consumes it via a new
`@classmoji/services/slides/ops` subpath (kept distinct from `…/slides` so
the deck tool tests' module mock leaves the real engine in place); the
starter-customCss drop on set_theme is now an option the caller passes.
Behavior is byte-identical — full MCP suite (305) green untouched.
