+++ packages/services/package.json
@@ -12,20 +12,22 @@
   "exports": {
     ".": "./src/index.ts",
     "./flags": "./src/classmoji/repoAnalytics.flags.ts",
-    "./slides": "./src/slides/index.ts"
+    "./slides": "./src/slides/index.ts",
+    "./slides/ops": "./src/slides/deckOps.ts"
   },
   "dependencies": {
     "@classmoji/database": "*",
     "@classmoji/utils": "*",
     "@octokit/auth-app": "^7.1.3",
-    "cheerio": "^1.1.2",
     "@trigger.dev/sdk": "4.4.0",
     "async": "^3.2.5",
+    "cheerio": "^1.1.2",
     "ics": "^3.8.1",
     "marked": "^15.0.4",
     "octokit": "^3.1.2",
     "prettier": "^3.6.2",
-    "simple-git": "^3.36.0"
+    "simple-git": "^3.36.0",
+    "zod": "^3.25.76"
   },
   "devDependencies": {
     "vitest": "^4.1.4"
